backend/pyproject.toml
# TODO: On-premise - Configure pip to use local artifactory by creating pip.conf or using --index-url
# Example pip.conf:
#   [global]
#   index-url = https://artifactory.your-company.com/api/pypi/pypi-remote/simple
#   trusted-host = artifactory.your-company.com

[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*"]

[tool.setuptools.package-data]
"core.api" = [
    "static/scalar/*.js",
    "static/scalar/*.svg",
    "static/scalar/fonts/*.woff2",
]

[project]
name = "skynet"
version = "0.1.0"
description = "Skynet - FastAPI service wrapper for DSPy optimization workflows"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.23",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
    "dspy~=3.0",
    "litellm>=1.40",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "pandas>=2.1",
    "openpyxl>=3.1",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "scalar-fastapi>=1.8",
    "scipy>=1.15.3",
    "fastmcp>=3.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.25",
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "ruff>=0.11",
]
load = [
    "locust>=2.0",
]
recommendations = [
    "sentence-transformers>=2.5",
    "numpy>=1.24",
]

[tool.ruff]
target-version = "py310"
line-length = 120
extend-exclude = ["usage_guide"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["core"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["RUF012"]
"core/**/tests/*" = ["RUF012"]
"core/notifications/*" = ["RUF001"]
"core/dataset/*" = ["RUF001"]

[tool.pytest.ini_options]
# Unit tests live next to the code they cover under ``core/<subpkg>/tests/``
# and run without Postgres, OPENAI_API_KEY, or a live server. Integration
# tests (test_llm_integration.py, test_load.py) live under ``tests/`` and
# require all three. Run each tier separately:
#
#   pytest core/ -v                          # fast CI gate, no credentials
#   pytest tests/test_llm_integration.py -v  # nightly paid tier
#   pytest tests/ -m smoke -v               # quick smoke pass (subset)
#   pytest tests/ -m "not slow" -v          # skip long-running LLM tests
#
testpaths = ["core", "tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = ["--strict-markers"]
markers = [
    "smoke: quick sanity checks — no LLM calls, just server availability and validation",
    "e2e: full end-to-end tests requiring a live LLM (slow, costs API credits)",
    "slow: tests that take > 30 s to complete",
    "load: concurrency / latency stress tests",
]
